Unload the form logic and move to plugins control component
